backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
openpyxl==3.1.2
gspread==5.12.0
google-auth==2.23.4
pydantic==2.9.2
python-multipart==0.0.6
python-dotenv==1.0.0
google-api-python-client==2.108.0
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
flask==3.0.0
twilio==8.10.0
requests==2.31.0
httpx==0.25.0
